add ppmod.detach, simplify ppmod.ontick
Added ppmod.detach - detaches a script function from the active control flow and pushes it to the end of the tick. If SQQuerySuspend is encountered, the function is run again recursively (args persist) until it finishes or until a different exception is caught. Useful for ensuring expensive loops get fully executed, yet equally dangerous due to the risk of freezing the game.

Simplified ppmod.ontick - the second argument, pause, sets whether the loop should stop on the pause screen. Defaults to true.